Acessos to int

+++ Classes/acessos.php
@@ -2,7 +2,7 @@
 
 class acesso{
     public function newAcesso(){
-        $acessos = file_get_contents('./acessos/acessos.txt');
+        $acessos =  (int) file_get_contents('./acessos/acessos.txt');
         $acessos++;
         file_put_contents("./acessos/acessos.txt", $acessos);
     }
